Fjern idporten endepunkt for personalia
