+++ src/main/kotlin/no/nav/tms/min/side/proxy/personalia/PersonaliaFetcher.kt
@@ -76,3 +76,55 @@ class PersonaliaFetcher(
     }
 }
 
+class HentNavnException(message: String, cause: Exception? = null): Exception(message, cause)
+
+class HentNavn(ident: String) {
+    val query = """
+        query(${'$'}ident: ID!) {
+            hentPerson(ident: ${'$'}ident) {
+                navn {
+                    fornavn,
+                    mellomnavn,
+                    etternavn
+                }
+            }
+        }
+    """.compactJson()
+
+    val variables = mapOf(
+        "ident" to ident
+    )
+}
+
+fun String.compactJson(): String =
+    trimIndent()
+        .replace("\r", " ")
+        .replace("\n", " ")
+        .replace("\\s+".toRegex(), " ")
+
+data class HentNavnResponse(
+    val data: HentNavnData?,
+    val errors: List<Map<String, Any>>?
+) {
+    data class HentNavnData(
+        val hentPerson: Person
+    )
+
+    data class Person (
+        val navn: List<Navn>
+    ) {
+        val fullnavn = navn.first().let {
+            listOf(it.fornavn, it.mellomnavn, it.etternavn)
+                .filter { navn -> !navn.isNullOrBlank() }
+                .joinToString(" ")
+        }
+    }
+
+    data class Navn(
+        val fornavn: String,
+        val mellomnavn: String? = null,
+        val etternavn: String,
+    )
+}
+
+
