+++ src/main/kotlin/no/nav/tms/min/side/proxy/personalia/personaliaRoutes.kt
@@ -17,5 +17,10 @@ fun Route.personaliaRoutes(personaliaFetcher: PersonaliaFetcher) {
     }
 }
 
+data class NavnAndIdent(
+    val navn: String?,
+    val ident: String
+)
+
 private val PipelineContext<Unit, ApplicationCall>.user
     get() = TokenXUserFactory.createTokenXUser(call)
