+++ src/main/kotlin/no/nav/tms/min/side/proxy/proxyApi.kt
@@ -33,7 +33,6 @@ fun Application.proxyApi(
     corsAllowedSchemes: String,
     contentFetcher: ContentFetcher,
     externalContentFetcher: ExternalContentFetcher,
-    navnFetcher: NavnFetcher,
     personaliaFetcher: PersonaliaFetcher,
     idportenAuthInstaller: Application.() -> Unit = {
         authentication {
@@ -120,7 +119,6 @@ fun Application.proxyApi(
             }
             proxyRoutes(contentFetcher, externalContentFetcher)
             aiaRoutes(externalContentFetcher)
-            navnRoutes(navnFetcher)
             get("featuretoggles") {
                 call.respond(
                     unleash.more().evaluateAllToggles().associate {
