+++ src/test/kotlin/no/nav/tms/min/side/proxy/GetRoutesTest.kt
@@ -49,7 +49,6 @@ class GetRoutesTest {
         mockApi(
             contentFetcher = contentFecther(proxyHttpClient),
             externalContentFetcher = externalContentFetcher(proxyHttpClient),
-            navnFetcher = mockk(),
             personaliaFetcher = mockk()
         )
 
@@ -97,7 +96,6 @@ class GetRoutesTest {
         mockApi(
             contentFetcher = contentFecther(proxyHttpClient),
             externalContentFetcher = externalContentFetcher(proxyHttpClient),
-            navnFetcher = mockk(),
             personaliaFetcher = mockk()
         )
 
@@ -131,7 +129,6 @@ class GetRoutesTest {
         mockApi(
             contentFetcher = contentFecther(proxyHttpClient),
             externalContentFetcher = externalContentFetcher(proxyHttpClient),
-            navnFetcher = mockk(),
             personaliaFetcher = mockk()
         )
 
@@ -151,7 +148,6 @@ class GetRoutesTest {
             contentFetcher = contentFecther(proxyHttpClient),
             externalContentFetcher = externalContentFetcher(proxyHttpClient),
             unleash = unleash,
-            navnFetcher = mockk(),
             personaliaFetcher = mockk()
         )
 
@@ -166,7 +162,6 @@ class GetRoutesTest {
         mockApi(
             contentFetcher = mockk(),
             externalContentFetcher = mockk(),
-            navnFetcher = mockk(),
             personaliaFetcher = mockk()
         )
         client.get("/authPing").status shouldBe HttpStatusCode.OK
@@ -178,7 +173,6 @@ class GetRoutesTest {
             mockApi(
                 contentFetcher = mockk(),
                 externalContentFetcher = mockk(),
-                navnFetcher = mockk(),
                 personaliaFetcher = mockk(),
                 levelOfAssurance = LevelOfAssurance.SUBSTANTIAL
             )
