+++ src/test/kotlin/no/nav/tms/min/side/proxy/PostRoutesTest.kt
@@ -21,7 +21,6 @@ class PostRoutesTest {
         mockApi(
             contentFetcher = contentFecther(proxyHttpClient),
             externalContentFetcher = externalContentFetcher(proxyHttpClient),
-            navnFetcher = mockk(),
             personaliaFetcher = mockk()
         )
